requirements.txt
Flask==3.0.3
requests==2.32.3
oracledb==3.3.0
Pillow==10.2.0